+++ test/carrier/action3_test.dart
@@ -7,8 +7,7 @@ class MockEvent3 extends Event3<String, String, String, String> {
   MockEvent3();
 
   @override
-  call(state, value1, value2, value3) =>
-      '$state $value1 $value2 $value3';
+  call(state, value1, value2, value3) => '$state $value1 $value2 $value3';
 }
 
 class MockStore extends Store<String> {
@@ -54,12 +53,9 @@ void main() {
       reducer1,
     );
     expect(objectUnterTest11.hashCode, objectUnterTest22.hashCode);
-    expect(objectUnterTest11.hashCode,
-        isNot(objectUnterTest12.hashCode));
-    expect(objectUnterTest11.hashCode,
-        isNot(objectUnterTest21.hashCode));
-    expect(objectUnterTest12.hashCode,
-        isNot(objectUnterTest21.hashCode));
+    expect(objectUnterTest11.hashCode, isNot(objectUnterTest12.hashCode));
+    expect(objectUnterTest11.hashCode, isNot(objectUnterTest21.hashCode));
+    expect(objectUnterTest12.hashCode, isNot(objectUnterTest21.hashCode));
   });
   test('Action3 operator== test', () {
     final store1 = MockStore('1');